refactor: refactor code and implement requested changes.
