+++ app/[network]/tokens/app.tsx
@@ -1,5 +1,6 @@
 "use client";
-import { ReactNode, useEffect } from "react";
+import { ReactNode } from "react";
+import { motion } from "framer-motion";
 
 import LBNavigation from "@/components/navigation";
 import { Network } from "@/components/button/types";
@@ -18,33 +19,12 @@ const networkBackgrounds: Record<Network, string> = {
 };
 
 const App = ({ children, params }: PageProps) => {
-  useEffect(() => {
-    const bgDiv = document.createElement("div");
-    bgDiv.style.position = "fixed";
-    bgDiv.style.top = "0";
-    bgDiv.style.left = "0";
-    bgDiv.style.width = "100vw";
-    bgDiv.style.height = "100vh";
-    bgDiv.style.backgroundImage = networkBackgrounds[params.network];
-    bgDiv.style.backgroundSize = "cover";
-    bgDiv.style.backgroundPosition = "center";
-    bgDiv.style.zIndex = "-10";
-    bgDiv.style.opacity = "0.1";
-    bgDiv.style.transition = "background 300ms ease-in-out";
-    bgDiv.style.pointerEvents = "none";
-
-    document.body.appendChild(bgDiv);
-
-    return () => {
-      if (document.body.contains(bgDiv)) {
-        document.body.removeChild(bgDiv);
-      }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [params.network]);
-
   return (
     <main>
+      <motion.div
+        animate={{ backgroundImage: networkBackgrounds[params.network] }}
+        className="fixed top-0 left-0 w-screen h-screen bg-cover bg-center -z-10 opacity-10 transition-all duration-300 ease-in-out pointer-events-none"
+      />
       <LBNavigation network={params.network} />
       {children}
       <div className="flex justify-center fixed w-screen bottom-0 -z-10">
