+++ app/type.ts
@@ -1,0 +1,8 @@
+import { Network } from "@/components/button/types";
+
+interface PageProps {
+  params: { network: Network };
+  children?: React.ReactNode;
+}
+
+export type { PageProps };
