+++ components/loader/types.ts
@@ -2,5 +2,3 @@ interface ISMLoader {
   variant?: "small" | "medium" | "large";
   color?: string;
 }
-
-export type { ISMLoader };
